+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented here.
 
+## Unreleased
+
+### Added
+
+- Remote SSH support with the companion extension and Pi running on the same workspace host.
+
+### Fixed
+
+- Resolve a live visible editor when active editor focus is lost, without caching selected text or retaining hidden/closed editors.
+
 ## [0.1.1] - 2026-07-19
 
 ## [0.1.1](https://github.com/IbrahimMohammed47/pi-vscode-context/compare/v0.1.0...v0.1.1) (2026-07-19)